broke down some large functions into smaller ones
